refactor: 자바 모듈 코틀린으로 변경 및 인터페이스화
