--- HIGOODS-Common/src/main/java/com/higoods/common/HigoodsCommonApplication.java
@@ -1,13 +1,0 @@
-package com.higoods.common;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class HigoodsCommonApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(HigoodsCommonApplication.class, args);
-    }
-
-}
