--- HIGOODS-Common/src/test/java/com/higoods/common/HigoodsCommonApplicationTests.java
@@ -1,13 +1,0 @@
-package com.higoods.common;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class HigoodsCommonApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
