+++ HIGOODS-Common/src/test/kotlin/com/higoods/common/HigoodsCommonApplicationTests.kt
@@ -1,0 +1,12 @@
+package com.higoods.common
+
+import org.junit.jupiter.api.Test
+import org.springframework.boot.test.context.SpringBootTest
+
+@SpringBootTest
+class HigoodsCommonApplicationTests {
+
+    @Test
+    fun contextLoads() {
+    }
+}
