--- HIGOODS-Domain/src/main/java/com/higoods/domain/HigoodsDomainApplication.java
@@ -1,13 +1,0 @@
-package com.higoods.domain;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class HigoodsDomainApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(HigoodsDomainApplication.class, args);
-    }
-
-}
