--- HIGOODS-Domain/src/test/java/com/higoods/domain/HigoodsDomainApplicationTests.java
@@ -1,13 +1,0 @@
-package com.higoods.domain;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class HigoodsDomainApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
