+++ HIGOODS-Domain/src/test/kotlin/com/higoods/domain/HigoodsDomainApplicationTests.kt
@@ -1,0 +1,12 @@
+package com.higoods.domain
+
+import org.junit.jupiter.api.Test
+import org.springframework.boot.test.context.SpringBootTest
+
+@SpringBootTest
+class HigoodsDomainApplicationTests {
+
+    @Test
+    fun contextLoads() {
+    }
+}
